Change dashboard edit to autoroute
